third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/dom-agent-query-selector-expected.txt

Tests DOMAgent.querySelector and DOMAgent.querySelectorAll.


Running: testDocumentQuerySelector
node id: id1

Running: testDocumentQuerySelectorAll
node id: id1
node id: id2
node id: id3
node id: id4
node id: id5
node id: id6

Running: testQuerySelector
node id: id3

Running: testQuerySelectorAll
node id: id3
node id: id4
node id: id5
node id: id6
